chapters/ch02/pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "blackhat-ai-ch02"
version = "0.1.0"
description = "Chapter 2: AI Agent Architecture - Companion code for Black Hat AI"
readme = "README.md"
requires-python = ">=3.9"
license = {text = "MIT"}
authors = [
    {name = "Black Hat AI Authors"}
]
keywords = ["ai", "security", "agents", "offensive-security", "red-team"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Education",
    "Topic :: Security",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]

dependencies = [
    "pydantic>=2.0.0",
    "python-dotenv>=1.0.0",
]

[project.optional-dependencies]
langchain = [
    "langchain>=0.1.0",
    "langchain-openai>=0.0.5",
    "openai>=1.0.0",
]
autogen = [
    "pyautogen>=0.2.0",
]
dev = [
    "pytest>=7.4.0",
    "pytest-cov>=4.1.0",
    "black>=23.0.0",
    "ruff>=0.1.0",
    "mypy>=1.7.0",
]
all = [
    "blackhat-ai-ch02[langchain,autogen,dev]",
]

[project.urls]
Homepage = "https://github.com/blackcat-labs/Black-Hat-AI"
Repository = "https://github.com/blackcat-labs/Black-Hat-AI"
Issues = "https://github.com/blackcat-labs/Black-Hat-AI/issues"

[tool.setuptools]
packages = ["src", "src.core", "src.tools", "src.adapters", "src.adapters.langchain", "src.adapters.autogen", "src.safety"]

[tool.black]
line-length = 100
target-version = ["py39", "py310", "py311", "py312"]
include = '\.pyi?$'
exclude = '''
/(
    \.git
  | \.venv
  | build
  | dist
  | __pycache__
)/
'''

[tool.ruff]
line-length = 100
target-version = "py39"
select = ["E", "F", "W", "I", "N", "UP"]
ignore = ["E501"]

[tool.mypy]
python_version = "3.9"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = false
ignore_missing_imports = true

[tool.pytest.ini_options]
minversion = "7.0"
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = "-v --cov=src --cov-report=term-missing"
